Refactoring StoneObstacleSession into a different file for better organization, timeouts now make the bot continue
